backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery==5.4.0
redis==5.0.8
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
pymysql==1.1.1
groq==0.9.0
langchain-groq==0.1.3
cohere
langgraph>=0.2,<0.3
llama-index-llms-openai
llama-index-retrievers-bm25
nest_asyncio==1.6.0
